lib: add error checking to BEGIN/COMMIT TRANSACTION statements

+++ exch/exmdb_provider/db_engine.cpp
@@ -1398,6 +1398,8 @@ static void db_engine_notify_content_table_add_row(db_item_ptr &pdb,
 				padded_row->after_row_id = inst_id1;
 			} else {
 				auto sql_transact = gx_sql_begin_trans(pdb->tables.psqlite);
+				if (!sql_transact)
+					continue;
 				snprintf(sql_string, arsizeof(sql_string), "UPDATE t%u SET idx=-(idx+1)"
 					" WHERE idx>=%u;UPDATE t%u SET idx=-idx WHERE"
 					" idx<0", ptable->table_id, idx, ptable->table_id);
@@ -1424,7 +1426,8 @@ static void db_engine_notify_content_table_add_row(db_item_ptr &pdb,
 				        " row_id=%llu", ptable->table_id, LLU{row_id}, LLU{row_id1});
 				if (gx_sql_exec(pdb->tables.psqlite, sql_string) != SQLITE_OK)
 					continue;
-				sql_transact.commit();
+				if (sql_transact.commit() != 0)
+					continue;
 				padded_row->after_row_id = inst_id;
 			}
 			if (ptable->table_flags & TABLE_FLAG_NONOTIFICATIONS)
@@ -1484,6 +1487,8 @@ static void db_engine_notify_content_table_add_row(db_item_ptr &pdb,
 			}
 		}
 		auto sql_transact = gx_sql_begin_trans(pdb->tables.psqlite);
+		if (!sql_transact)
+			continue;
 		char sql_string[164];
 		snprintf(sql_string, arsizeof(sql_string), "SELECT row_id, inst_id, "
 		         "value FROM t%u WHERE prev_id=?", ptable->table_id);
@@ -1755,7 +1760,8 @@ static void db_engine_notify_content_table_add_row(db_item_ptr &pdb,
 			return;
 		pstmt.finalize();
 		pstmt1.finalize();
-		sql_transact.commit();
+		if (sql_transact.commit() != 0)
+			continue;
 		if (ptable->table_flags & TABLE_FLAG_NONOTIFICATIONS)
 			continue;
 		if (b_resorted) {
@@ -2061,12 +2067,15 @@ static void db_engine_notify_hierarchy_table_add_row(db_item_ptr &pdb,
 			if (idx == 0)
 				goto APPEND_END_OF_TABLE;
 			auto sql_transact = gx_sql_begin_trans(pdb->tables.psqlite);
+			if (!sql_transact)
+				continue;
 			snprintf(sql_string, arsizeof(sql_string), "UPDATE t%u SET idx=-(idx+1)"
 				" WHERE idx>%u;UPDATE t%u SET idx=-idx WHERE"
 				" idx<0", ptable->table_id, idx, ptable->table_id);
 			if (gx_sql_exec(pdb->tables.psqlite, sql_string) != SQLITE_OK)
 				continue;
-			sql_transact.commit();
+			if (sql_transact.commit() != 0)
+				continue;
 			snprintf(sql_string, arsizeof(sql_string), "INSERT INTO t%u (idx, "
 				"folder_id, depth) VALUES (%u, %llu, %u)",
 				ptable->table_id, idx + 1, LLU{folder_id}, depth);
@@ -2278,6 +2287,8 @@ static void db_engine_notify_content_table_delete_row(db_item_ptr &pdb,
 			prev_id = sqlite3_column_int64(pstmt, 2);
 			pstmt.finalize();
 			auto sql_transact = gx_sql_begin_trans(pdb->tables.psqlite);
+			if (!sql_transact)
+				continue;
 			snprintf(sql_string, arsizeof(sql_string), "DELETE FROM t%u WHERE "
 				"row_id=%llu", ptable->table_id, LLU{row_id});
 			if (gx_sql_exec(pdb->tables.psqlite, sql_string) != SQLITE_OK)
@@ -2296,7 +2307,8 @@ static void db_engine_notify_content_table_delete_row(db_item_ptr &pdb,
 				ptable->table_id, ptable->table_id);
 			if (gx_sql_exec(pdb->tables.psqlite, sql_string) != SQLITE_OK)
 				continue;
-			sql_transact.commit();
+			if (sql_transact.commit() != 0)
+				continue;
 			if (ptable->table_flags & TABLE_FLAG_NONOTIFICATIONS)
 				continue;
 			if (!common_util_get_message_parent_folder(pdb->psqlite,
@@ -2345,6 +2357,8 @@ static void db_engine_notify_content_table_delete_row(db_item_ptr &pdb,
 		}
 		pstmt.finalize();
 		auto sql_transact = gx_sql_begin_trans(pdb->tables.psqlite);
+		if (!sql_transact)
+			continue;
 		snprintf(sql_string, arsizeof(sql_string), "SELECT * FROM"
 			" t%u WHERE row_id=?", ptable->table_id);
 		pstmt = gx_sql_prep(pdb->tables.psqlite, sql_string);
@@ -2557,7 +2571,8 @@ static void db_engine_notify_content_table_delete_row(db_item_ptr &pdb,
 			pstmt.finalize();
 			pstmt1.finalize();
 		}
-		sql_transact.commit();
+		if (sql_transact.commit() != 0)
+			continue;
 		if (ptable->table_flags & TABLE_FLAG_NONOTIFICATIONS)
 			continue;
 		if (b_resorted) {
